63182: Avoid extra notifications when using non container threads on read causing thread safety problems. Tentative fix, but the pattern is clearly causing a thread safety problem there. Also move ContainerMarkerThread, but leave the old location around for compatibility.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1853886 13f79535-47bb-0310-9956-ffa450edef68
