+++ java/org/apache/tomcat/util/net/Nio2Endpoint.java
@@ -811,7 +811,7 @@ public class Nio2Endpoint extends AbstractJsseEndpoint<Nio2Channel,AsynchronousS
                     socketBufferHandler.configureReadBufferForRead();
                     nRead = Math.min(nRead, len);
                     socketBufferHandler.getReadBuffer().get(b, off, nRead);
-                } else if (nRead == 0 && !block) {
+                } else if (nRead == 0 && !block && ContainerThreadMarker.isContainerThread()) {
                     readInterest = true;
                 }
 
@@ -876,7 +876,7 @@ public class Nio2Endpoint extends AbstractJsseEndpoint<Nio2Channel,AsynchronousS
                     // data that was just read
                     if (nRead > 0) {
                         nRead = populateReadBuffer(to);
-                    } else if (nRead == 0 && !block) {
+                    } else if (nRead == 0 && !block && ContainerThreadMarker.isContainerThread()) {
                         readInterest = true;
                     }
                 }
