+++ test/org/apache/catalina/nonblocking/TestNonBlockingAPI.java
@@ -48,12 +48,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.junit.Assert;
-import org.junit.Assume;
 import org.junit.Ignore;
 import org.junit.Test;
 
 import org.apache.catalina.Context;
-import org.apache.catalina.connector.Connector;
 import org.apache.catalina.startup.BytesStreamer;
 import org.apache.catalina.startup.TesterServlet;
 import org.apache.catalina.startup.Tomcat;
@@ -95,11 +93,6 @@ public class TestNonBlockingAPI extends TomcatBaseTest {
 
     @Test
     public void testNonBlockingReadAsync() throws Exception {
-        Tomcat tomcat = getTomcatInstance();
-        Connector connector = tomcat.getConnector();
-        // Skip for NIO2
-        Assume.assumeFalse("This test may fail for NIO2",
-                connector.getProtocolHandlerClassName().contains("Nio2"));
         doTestNonBlockingRead(false, true);
     }
 
@@ -124,7 +117,7 @@ public class TestNonBlockingAPI extends TomcatBaseTest {
         tomcat.start();
 
         Map<String, List<String>> resHeaders = new HashMap<>();
-        int rc = postUrl(true, new DataWriter(async ? 0 : 500, async ? 1000000 : 5),
+        int rc = postUrl(true, new DataWriter(async ? 0 : 500, async ? 2000000 : 5),
                 "http://localhost:" + getPort() + "/", new ByteChunk(), resHeaders, null);
 
         Assert.assertEquals(HttpServletResponse.SC_OK, rc);
